pyproject.toml
[tool.poetry]
name = "simple-threads"
version = "0.1.0"
description = ""
authors = ["Memo Chou <memochou1993@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.2"
mangum = "^0.17.0"
boto3 = "^1.34.94"
uuid = "^1.30"
email-validator = "^2.1.1"

[tool.poetry.group.dev.dependencies]
aws-cdk-lib = "^2.139.0"
poetry-plugin-export = "^1.7.1"
pytest = "^8.1.2"
python-dotenv = "^1.0.1"
ruff = "^0.4.2"
uvicorn = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
